first pass - move search code to a search package
