+++ internal/search/windows.go
@@ -1,4 +1,4 @@
-package command
+package search
 
 import (
 	"bytes"
