+++ pkg/coderefs/coderefs.go
@@ -1,55 +1,33 @@
 package coderefs
 
 import (
-	"container/list"
 	"fmt"
 	"os"
-	"regexp"
 	"sort"
 	"strings"
 	"time"
 
-	"github.com/launchdarkly/ld-find-code-refs/internal/command"
+	"github.com/launchdarkly/ld-find-code-refs/internal/git"
 	"github.com/launchdarkly/ld-find-code-refs/internal/helpers"
 	"github.com/launchdarkly/ld-find-code-refs/internal/ld"
 	"github.com/launchdarkly/ld-find-code-refs/internal/log"
 	"github.com/launchdarkly/ld-find-code-refs/internal/options"
+	"github.com/launchdarkly/ld-find-code-refs/internal/search"
 	"github.com/launchdarkly/ld-find-code-refs/internal/validation"
 	"github.com/launchdarkly/ld-find-code-refs/internal/version"
 )
 
-// These are defensive limits intended to prevent corner cases stemming from
-// large repos, false positives, etc. The goal is a) to prevent the program
-// from taking a very long time to run and b) to prevent the program from
-// PUTing a massive json payload. These limits will likely be tweaked over
-// time. The LaunchDarkly backend will also apply limits.
 const (
-	minFlagKeyLen                     = 3     // Minimum flag key length helps reduce the number of false positives
-	maxFileCount                      = 10000 // Maximum number of files containing code references
-	maxLineCharCount                  = 500   // Maximum number of characters per line
-	maxHunkCount                      = 25000 // Maximum number of total code references
-	maxHunkedLinesPerFileAndFlagCount = 500   // Maximum number of lines per flag in a file
-	maxProjKeyLength                  = 20    // Maximum project key length
+	minFlagKeyLen    = 3  // Minimum flag key length helps reduce the number of false positives
+	maxProjKeyLength = 20 // Maximum project key length
 )
-
-// map of flag keys to slices of lines those flags occur on
-type flagReferenceMap map[string][]*list.Element
-
-// this struct contains a linked list of all the search result lines
-// for a single file, and a map of flag keys to slices of lines where
-// those flags occur.
-type fileSearchResults struct {
-	path                  string
-	fileSearchResultLines *list.List
-	flagReferenceMap
-}
 
 type branch struct {
 	Name             string
 	Head             string
 	UpdateSequenceId *int
 	SyncTime         int64
-	SearchResults    searchResultLines
+	SearchResults    search.SearchResultLines
 }
 
 func Scan(opts options.Options) {
@@ -65,12 +43,12 @@ func Scan(opts options.Options) {
 	}
 
 	log.Info.Printf("absolute directory path: %s", absPath)
-	searchClient, err := command.NewAgClient(absPath)
+	searchClient, err := search.NewAgClient(absPath)
 	if err != nil {
 		log.Error.Fatalf("%s", err)
 	}
 
-	gitClient, err := command.NewGitClient(absPath, opts.Branch)
+	gitClient, err := git.NewClient(absPath, opts.Branch)
 	if err != nil {
 		log.Error.Fatalf("%s", err)
 	}
@@ -151,7 +129,7 @@ func Scan(opts options.Options) {
 	delims = append(delims, opts.Delimiters.Additional...)
 	delimString := strings.Join(helpers.Dedupe(delims), "")
 
-	refs, err := findReferences(searchClient, filteredFlags, aliases, ctxLines, delimString)
+	refs, err := searchClient.FindReferences(filteredFlags, aliases, ctxLines, delimString)
 	if err != nil {
 		log.Error.Fatalf("error searching for flag key references: %s", err)
 	}
@@ -268,262 +246,18 @@ func getFlags(ldApi ld.ApiClient) ([]string, error) {
 	return flags, nil
 }
 
-func findReferencedFlags(ref string, aliases map[string][]string, delims string) map[string][]string {
-	ret := make(map[string][]string, len(aliases))
-	for key, flagAliases := range aliases {
-		matcher := regexp.MustCompile(regexp.QuoteMeta(key))
-		if len(delims) > 0 {
-			matcher = regexp.MustCompile(fmt.Sprintf("[%s]%s[%s]", delims, regexp.QuoteMeta(key), delims))
-		}
-		if matcher.MatchString(ref) {
-			ret[key] = make([]string, 0, len(flagAliases))
-		}
-		for _, alias := range flagAliases {
-			aliasMatcher := regexp.MustCompile(regexp.QuoteMeta(alias))
-			if aliasMatcher.MatchString(ref) {
-				if ret[key] == nil {
-					ret[key] = make([]string, 0, len(flagAliases))
-				}
-				ret[key] = append(ret[key], alias)
-			}
-		}
-	}
-	return ret
-}
-
 func (b *branch) makeBranchRep(projKey string, ctxLines int) ld.BranchRep {
 	return ld.BranchRep{
 		Name:             strings.TrimPrefix(b.Name, "refs/heads/"),
 		Head:             b.Head,
 		UpdateSequenceId: b.UpdateSequenceId,
 		SyncTime:         b.SyncTime,
-		References:       b.SearchResults.makeReferenceHunksReps(projKey, ctxLines),
-	}
-}
-
-func (g searchResultLines) makeReferenceHunksReps(projKey string, ctxLines int) []ld.ReferenceHunksRep {
-	reps := []ld.ReferenceHunksRep{}
-
-	aggregatedSearchResults := g.aggregateByPath()
-
-	if len(aggregatedSearchResults) > maxFileCount {
-		log.Warning.Printf("found %d files with code references, which exceeded the limit of %d", len(aggregatedSearchResults), maxFileCount)
-		aggregatedSearchResults = aggregatedSearchResults[0:maxFileCount]
-	}
-
-	numHunks := 0
-
-	shouldSuppressUnexpectedError := false
-	for _, fileSearchResults := range aggregatedSearchResults {
-		if numHunks > maxHunkCount {
-			log.Warning.Printf("found %d code references across all files, which exceeeded the limit of %d. halting code reference search", numHunks, maxHunkCount)
-			break
-		}
-
-		hunks := fileSearchResults.makeHunkReps(projKey, ctxLines)
-
-		if len(hunks) == 0 && !shouldSuppressUnexpectedError {
-			log.Error.Printf("expected code references but found none in '%s'", fileSearchResults.path)
-			log.Debug.Printf("%+v", fileSearchResults)
-			// if this error occurred, it's likely to occur for many other files, and create a lot of noise. So, suppress the message for all other occurrences
-			shouldSuppressUnexpectedError = true
-			continue
-		}
-
-		numHunks += len(hunks)
-
-		reps = append(reps, ld.ReferenceHunksRep{Path: fileSearchResults.path, Hunks: hunks})
-	}
-	return reps
-}
-
-// Assumes invariant: searchResultLines will already be sorted by path.
-func (g searchResultLines) aggregateByPath() []fileSearchResults {
-	allFileResults := []fileSearchResults{}
-
-	if len(g) == 0 {
-		return allFileResults
-	}
-
-	// initialize first file
-	currentFileResults := fileSearchResults{
-		path:                  g[0].Path,
-		flagReferenceMap:      flagReferenceMap{},
-		fileSearchResultLines: list.New(),
-	}
-
-	for _, searchResult := range g {
-		// If we reach a search result with a new path, append the old one to our list and start a new one
-		if searchResult.Path != currentFileResults.path {
-			allFileResults = append(allFileResults, currentFileResults)
-
-			currentFileResults = fileSearchResults{
-				path:                  searchResult.Path,
-				flagReferenceMap:      flagReferenceMap{},
-				fileSearchResultLines: list.New(),
-			}
-		}
-
-		elem := currentFileResults.addSearchResult(searchResult)
-
-		if len(searchResult.FlagKeys) > 0 {
-			for flagKey := range searchResult.FlagKeys {
-				currentFileResults.addFlagReference(flagKey, elem)
-			}
-		}
-	}
-
-	// append last file
-	allFileResults = append(allFileResults, currentFileResults)
-
-	return allFileResults
-}
-
-func (fsr *fileSearchResults) addSearchResult(searchResult searchResultLine) *list.Element {
-	prev := fsr.fileSearchResultLines.Back()
-	if prev != nil && prev.Value.(searchResultLine).LineNum > searchResult.LineNum {
-		// This should never happen, as `ag` (and any other search program we might use
-		// should always return search results sorted by line number. We sanity check
-		// that lines are sorted _just in case_ since the downstream hunking algorithm
-		// only works on sorted lines.
-		log.Error.Fatalf("search results returned out of order")
-	}
-
-	return fsr.fileSearchResultLines.PushBack(searchResult)
-}
-
-func (fsr *fileSearchResults) addFlagReference(key string, ref *list.Element) {
-	_, ok := fsr.flagReferenceMap[key]
-
-	if ok {
-		fsr.flagReferenceMap[key] = append(fsr.flagReferenceMap[key], ref)
-	} else {
-		fsr.flagReferenceMap[key] = []*list.Element{ref}
-	}
-}
-
-func (fsr fileSearchResults) makeHunkReps(projKey string, ctxLines int) []ld.HunkRep {
-	hunks := []ld.HunkRep{}
-
-	for flagKey, flagReferences := range fsr.flagReferenceMap {
-		flagHunks := buildHunksForFlag(projKey, flagKey, fsr.path, flagReferences, ctxLines)
-		hunks = append(hunks, flagHunks...)
-	}
-
-	return hunks
-}
-
-func buildHunksForFlag(projKey, flag, path string, flagReferences []*list.Element, ctxLines int) []ld.HunkRep {
-	hunks := []ld.HunkRep{}
-
-	var previousHunk *ld.HunkRep
-	var currentHunk ld.HunkRep
-
-	lastSeenLineNum := -1
-
-	var hunkStringBuilder strings.Builder
-
-	appendToPreviousHunk := false
-
-	numHunkedLines := 0
-
-	for _, ref := range flagReferences {
-		// Each ref is either the start of a new hunk or a continuation of the previous hunk.
-		// NOTE: its possible that this flag reference is totally contained in the previous hunk
-		ptr := ref
-		numCtxLinesBeforeFlagRef := 0
-
-		// Attempt to seek to the start of the new hunk.
-		for i := 0; i < ctxLines; i++ {
-			// If we seek to a nil pointer, we're at the start of the file and can go no further.
-			if ptr.Prev() != nil {
-				ptr = ptr.Prev()
-				numCtxLinesBeforeFlagRef++
-			}
-
-			// If we seek earlier than the end of the last hunk, this reference overlaps at least
-			// partially with the last hunk and we should (possibly) expand the previous hunk rather than
-			// starting a new hunk.
-			if ptr.Value.(searchResultLine).LineNum <= lastSeenLineNum {
-				appendToPreviousHunk = true
-			}
-		}
-
-		// If we are starting a new hunk, initialize it
-		if !appendToPreviousHunk {
-			currentHunk = initHunk(projKey, flag)
-			currentHunk.StartingLineNumber = ptr.Value.(searchResultLine).LineNum
-			hunkStringBuilder.Reset()
-		}
-
-		// From the current position (at the theoretical start of the hunk) seek forward line by line X times,
-		// where X = (numCtxLinesBeforeFlagRef + 1 + ctxLines). Note that if the flag reference occurs close to the
-		// start of the file, numCtxLines may be smaller than ctxLines.
-		//   For each line, check if we have seeked past the end of the last hunk
-		//     If so: write that line to the hunkStringBuilder
-		//     Record that line as the last seen line.
-		for i := 0; i < numCtxLinesBeforeFlagRef+1+ctxLines; i++ {
-			ptrLineNum := ptr.Value.(searchResultLine).LineNum
-			if ptrLineNum > lastSeenLineNum {
-				lineText := truncateLine(ptr.Value.(searchResultLine).LineText)
-				hunkStringBuilder.WriteString(lineText + "\n")
-				lastSeenLineNum = ptrLineNum
-				numHunkedLines += 1
-				currentHunk.Aliases = append(currentHunk.Aliases, ptr.Value.(searchResultLine).FlagKeys[flag]...)
-			}
-
-			if ptr.Next() != nil {
-				ptr = ptr.Next()
-			}
-		}
-
-		if appendToPreviousHunk {
-			previousHunk.Lines = hunkStringBuilder.String()
-			appendToPreviousHunk = false
-		} else {
-			currentHunk.Lines = hunkStringBuilder.String()
-			currentHunk.Aliases = helpers.Dedupe(currentHunk.Aliases)
-			hunks = append(hunks, currentHunk)
-			previousHunk = &hunks[len(hunks)-1]
-		}
-
-		// If we have written more than the max. allowed number of lines for this file and flag, finish this hunk and exit early.
-		// This guards against a situation where the user has very long files with many false positive matches.
-		if numHunkedLines > maxHunkedLinesPerFileAndFlagCount {
-			log.Warning.Printf("found %d code reference lines in %s for the flag %s, which exceeded the limit of %d. truncating code references for this path and flag.",
-				numHunkedLines, path, flag, maxHunkedLinesPerFileAndFlagCount)
-			return hunks
-		}
-	}
-
-	return hunks
-}
-
-func initHunk(projKey, flagKey string) ld.HunkRep {
-	return ld.HunkRep{
-		ProjKey: projKey,
-		FlagKey: flagKey,
-		Aliases: []string{},
+		References:       b.SearchResults.MakeReferenceHunksReps(projKey, ctxLines),
 	}
 }
 
 func makeTimestamp() int64 {
 	return time.Now().UnixNano() / int64(time.Millisecond)
-}
-
-// Truncate lines to prevent sending over massive hunks, e.g. a minified file.
-// NOTE: We may end up truncating a valid flag key reference. We accept this risk
-//       and will handle hunks missing flag key references on the frontend.
-func truncateLine(line string) string {
-	// len(line) returns number of bytes, not num. characters, but it's a close enough
-	// approximation for our purposes
-	if len(line) > maxLineCharCount {
-		// convert to rune slice so that we don't truncate multibyte unicode characters
-		runes := []rune(line)
-		return string(runes[0:maxLineCharCount]) + "…"
-	} else {
-		return line
-	}
 }
 
 func fatalServiceError(err error, ignoreServiceErrors bool) {
